refactor vc integration tests into unit tests in ion/attestor and api/api, add appropriate test ignore tags
