+++ trustchain-ion/src/attestor.rs
@@ -171,6 +171,14 @@ impl Issuer for IONAttestor {
 
 #[async_trait]
 impl Holder for IONAttestor {
+    // This implementation ensures that the holder field is set on the Presentation, with the
+    // following implications:
+    //   - proof generation is handled by the ssi library
+    //   - ssi::ldp::ensure_or_pick_verification_relationship calls presentation.get_issuer()
+    //      which returns the holder (if Some, which is always the case)
+    //   - ensure_or_pick_verification_relationship tries to resolve the holder DID and check it's
+    //      verification methods
+    //   - so the holder's DID must be resolvable
     async fn sign_presentation<T: DIDResolver>(
         &self,
         presentation: &Presentation,
@@ -309,6 +317,45 @@ mod tests {
         assert!(vc_with_proof.is_ok());
     }
 
+    #[ignore = "requires a running Sidetree node listening on http://localhost:3000"]
+    #[tokio::test]
+    async fn test_sign_credential_failure() {
+        // Initialize temp path for saving keys
+        init();
+
+        // 1. Set-up (with a DID that will *not* match the issuer field in the credential).
+        let did = "did:ion:test:EiDMe2SFfJ_7eXVW7RF1ZHOkeu2M-Bre0ak2cXNBH0P-TQ";
+
+        // Make resolver
+        let resolver = get_ion_resolver("http://localhost:3000/");
+
+        // 2. Load Attestor
+        // let attestor = IONAttestor::new(did);
+        // Attestor
+        let attestor = IONAttestor::try_from(AttestorData::new(
+            did.to_string(),
+            serde_json::from_str(TEST_SIGNING_KEYS).unwrap(),
+        ))
+        .unwrap();
+
+        // 3. Read credential and set issuer field
+        let mut vc: Credential = serde_json::from_str(TEST_CREDENTIAL).unwrap();
+        vc.issuer = Some(ssi::vc::Issuer::URI(URI::String(
+            "did:ion:test:EiAtHHKFJWAk5AsM3tgCut3OiBY4ekHTf66AAjoysXL65Q".to_string(),
+        )));
+
+        // Sign credential (expect failure).
+        // Note: Signing a vc with a Some() issuer field requires a running ion node
+        let vc_with_proof = attestor.sign(&vc, None, None, &resolver).await;
+        assert!(vc_with_proof.is_err());
+
+        assert!(matches!(
+            vc_with_proof,
+            Err(IssuerError::SSI(ssi::error::Error::KeyMismatch))
+        ));
+    }
+
+    #[ignore = "requires a running Sidetree node listening on http://localhost:3000"]
     #[tokio::test]
     async fn test_attest_presentation() {
         // Note: removed tmp directory overwrite for TRUSTCHAIN_DATA, to have access to
@@ -327,29 +374,23 @@ mod tests {
         //     serde_json::from_str(TEST_SIGNING_KEYS).unwrap(),
         // ))
         // .unwrap();
+
         let vc = serde_json::from_str(TEST_CREDENTIAL).unwrap();
         let vc_with_proof = issuer.sign(&vc, None, None, &resolver).await.unwrap();
+
+        // Create Presentation, initially with holder field defaulting to None
         let presentation = Presentation {
             verifiable_credential: Some(OneOrMany::One(CredentialOrJWT::Credential(vc_with_proof))),
             ..Default::default()
         };
-        // assert holder field is not initially populated
-        assert!(presentation.holder.is_none());
 
-        // Attest to vp:
-        // .sign_presenatation now checks and sets the holder field of the presentation
-        // This has implications for the proof generation which is handled by the ssi library:
-        //   - ssi::ldp::ensure_or_pick_verification_relationship calls presentation.get_issuer()
-        //      which returns the holder (if Some)
-        //   - ensure_or_pick_verification_relationship tries to resolve the did and check it's
-        //      verification methods
+        // Holder field set to the DID of the signing holder by 'sign_presentation'
+        // The DID is resolved during signing, which requires a running ion node.
         let vp = holder
             .sign_presentation(&presentation, None, &resolver, None)
             .await;
-        assert!(vp.is_ok());
 
-        // Check holder field has been correctly populated during signing
-        assert_eq!(vp.unwrap().holder.unwrap().to_string(), holder.did);
+        assert!(vp.is_ok());
     }
 
     #[test]
